重构数据类型转换逻辑
对数据类型转换的实现进行了全面重构和优化，主要包括：
- 移除 `Valid.Boolean.cs` 中的 `TRUE_STRINGS` 定义，简化 `ToBoolean` 方法。
- 在 `Valid.Byte.cs` 和其他文件中添加了从基本数据类型到布尔值及其他类型的转换方法，增强了灵活性和一致性。
- 引入了格式化字符串支持，特别是在 `Valid.DateTime.cs` 中。
- 增强了对可空类型的支持，允许将 `null` 值转换为默认值。
- 提升了代码的可读性和可维护性，减少了重复代码。
