+++ src/LuYao.Common/Valid/Valid.DateTime.cs
@@ -5,86 +5,59 @@ namespace LuYao;
 
 partial class Valid
 {
-    /// <inheritdoc/>
-    public static DateTime ToDateTime(long value) => DateTime.FromBinary(value);
 
-    /// <inheritdoc/>
-    public static DateTime ToDateTime(string? value)
+    /// <summary>
+    /// 将值转换为字符串。
+    /// </summary>
+    public static string ToString(DateTime dt, Date format)
     {
-        if (string.IsNullOrWhiteSpace(value)) return default;
-        if (DateTime.TryParse(value, out var dt)) return dt;
-        return default;
+        if (dt == default(DateTime)) return string.Empty;
+        switch (format)
+        {
+            case Date.月日: return dt.ToString("MM/dd", DateTimeFormatInfo.InvariantInfo);
+            case Date.小时分钟秒: return dt.ToString("HH:mm:ss", DateTimeFormatInfo.InvariantInfo);
+            case Date.小时分钟: return dt.ToString("HH:mm", DateTimeFormatInfo.InvariantInfo);
+            case Date.年月日小时分钟: return dt.ToString("yyyy/MM/dd HH:mm", DateTimeFormatInfo.InvariantInfo);
+            case Date.月日年小时分钟: return dt.ToString("MM/dd/yyyy HH:mm", DateTimeFormatInfo.InvariantInfo);
+            case Date.月日小时分钟: return dt.ToString("MM/dd HH:mm", DateTimeFormatInfo.InvariantInfo);
+            case Date.全部显示: return dt.ToString("yyyy/MM/dd HH:mm:ss", DateTimeFormatInfo.InvariantInfo);
+            default: return dt.ToString("yyyy/MM/dd", DateTimeFormatInfo.InvariantInfo);
+        }
     }
 
-    /// <inheritdoc/>
-    public static DateTime ToDateTime(bool value) => default;
+    /// <summary>
+    /// 将值转换为字符串。
+    /// </summary>
+    public static string ToString(DateTime? value, Date format) => ToString(value ?? DateTime.MinValue, format);
 
     /// <inheritdoc/>
-    public static DateTime ToDateTime(char value) => default;
-
+    public static string ToString(DateTime value) => value.ToString("O");
     /// <inheritdoc/>
-    public static DateTime ToDateTime(sbyte value) => default;
-
+    public static Boolean ToBoolean(DateTime value) => Convert.ToBoolean(value);
     /// <inheritdoc/>
-    public static DateTime ToDateTime(byte value) => default;
-
+    public static Char ToChar(DateTime value) => Convert.ToChar(value);
     /// <inheritdoc/>
-    public static DateTime ToDateTime(short value) => default;
-
+    public static SByte ToSByte(DateTime value) => Convert.ToSByte(value);
     /// <inheritdoc/>
-    public static DateTime ToDateTime(ushort value) => default;
-
+    public static Byte ToByte(DateTime value) => Convert.ToByte(value);
     /// <inheritdoc/>
-    public static DateTime ToDateTime(int value)
-    {
-        long ticks = value;
-        if (ticks < DateTime.MinValue.Ticks || ticks > DateTime.MaxValue.Ticks) return default;
-        return new DateTime(ticks);
-    }
-
+    public static Int16 ToInt16(DateTime value) => Convert.ToInt16(value);
     /// <inheritdoc/>
-    public static DateTime ToDateTime(uint value)
-    {
-        long ticks = value;
-        if (ticks < DateTime.MinValue.Ticks || ticks > DateTime.MaxValue.Ticks) return default;
-        return new DateTime(ticks);
-    }
-
+    public static UInt16 ToUInt16(DateTime value) => Convert.ToUInt16(value);
     /// <inheritdoc/>
-    public static DateTime ToDateTime(ulong value)
-    {
-        // ulong 可能超出 long 范围
-        if (value > (ulong)DateTime.MaxValue.Ticks) return default;
-        long ticks = (long)value;
-        if (ticks < DateTime.MinValue.Ticks || ticks > DateTime.MaxValue.Ticks) return default;
-        return new DateTime(ticks);
-    }
-
+    public static Int32 ToInt32(DateTime value) => Convert.ToInt32(value);
     /// <inheritdoc/>
-    public static DateTime ToDateTime(float value)
-    {
-        // System.Convert.ToDateTime(float) 先转 double，再转 ticks
-        if (float.IsNaN(value) || float.IsInfinity(value)) return default;
-        double d = value;
-        if (d < DateTime.MinValue.Ticks || d > DateTime.MaxValue.Ticks) return default;
-        return new DateTime((long)d);
-    }
-
+    public static UInt32 ToUInt32(DateTime value) => Convert.ToUInt32(value);
     /// <inheritdoc/>
-    public static DateTime ToDateTime(double value)
-    {
-        if (double.IsNaN(value) || double.IsInfinity(value)) return default;
-        if (value < DateTime.MinValue.Ticks || value > DateTime.MaxValue.Ticks) return default;
-        return new DateTime((long)value);
-    }
-
+    public static Int64 ToInt64(DateTime value) => Convert.ToInt64(value);
     /// <inheritdoc/>
-    public static DateTime ToDateTime(decimal value)
-    {
-        if (value < DateTime.MinValue.Ticks || value > DateTime.MaxValue.Ticks) return default;
-        return new DateTime((long)value);
-    }
-
+    public static UInt64 ToUInt64(DateTime value) => Convert.ToUInt64(value);
+    /// <inheritdoc/>
+    public static Single ToSingle(DateTime value) => Convert.ToSingle(value);
+    /// <inheritdoc/>
+    public static Double ToDouble(DateTime value) => Convert.ToDouble(value);
+    /// <inheritdoc/>
+    public static Decimal ToDecimal(DateTime value) => Convert.ToDecimal(value);
     /// <inheritdoc/>
     public static DateTime ToDateTime(DateTime value) => value;
 }
