+++ src/LuYao.Common/Valid/Valid.Decimal.cs
@@ -5,31 +5,33 @@ namespace LuYao;
 partial class Valid
 {
     /// <inheritdoc/>
-    public static Decimal ToDecimal(Boolean value) => value ? 1 : 0;
+    public static Boolean ToBoolean(Decimal value) => Convert.ToBoolean(value);
     /// <inheritdoc/>
-    public static Decimal ToDecimal(Char value) => Convert.ToDecimal(value);
+    public static Char ToChar(Decimal value) => Convert.ToChar(value);
     /// <inheritdoc/>
-    public static Decimal ToDecimal(SByte value) => Convert.ToDecimal(value);
+    public static SByte ToSByte(Decimal value) => Convert.ToSByte(value);
     /// <inheritdoc/>
-    public static Decimal ToDecimal(Byte value) => Convert.ToDecimal(value);
+    public static Byte ToByte(Decimal value) => Convert.ToByte(value);
     /// <inheritdoc/>
-    public static Decimal ToDecimal(Int16 value) => Convert.ToDecimal(value);
+    public static Int16 ToInt16(Decimal value) => Convert.ToInt16(value);
     /// <inheritdoc/>
-    public static Decimal ToDecimal(UInt16 value) => Convert.ToDecimal(value);
+    public static UInt16 ToUInt16(Decimal value) => Convert.ToUInt16(value);
     /// <inheritdoc/>
-    public static Decimal ToDecimal(Int32 value) => Convert.ToDecimal(value);
+    public static Int32 ToInt32(Decimal value) => Convert.ToInt32(value);
     /// <inheritdoc/>
-    public static Decimal ToDecimal(UInt32 value) => Convert.ToDecimal(value);
+    public static UInt32 ToUInt32(Decimal value) => Convert.ToUInt32(value);
     /// <inheritdoc/>
-    public static Decimal ToDecimal(Int64 value) => Convert.ToDecimal(value);
+    public static Int64 ToInt64(Decimal value) => Convert.ToInt64(value);
     /// <inheritdoc/>
-    public static Decimal ToDecimal(UInt64 value) => Convert.ToDecimal(value);
+    public static UInt64 ToUInt64(Decimal value) => Convert.ToUInt64(value);
     /// <inheritdoc/>
-    public static Decimal ToDecimal(Single value) => Convert.ToDecimal(value);
+    public static Single ToSingle(Decimal value) => Convert.ToSingle(value);
     /// <inheritdoc/>
-    public static Decimal ToDecimal(Double value) => Convert.ToDecimal(value);
+    public static Double ToDouble(Decimal value) => Convert.ToDouble(value);
     /// <inheritdoc/>
     public static Decimal ToDecimal(Decimal value) => value;
     /// <inheritdoc/>
-    public static Decimal ToDecimal(DateTime value) => Convert.ToDecimal(value);
+    public static DateTime ToDateTime(Decimal value) => Convert.ToDateTime(value);
+    /// <inheritdoc/>
+    public static String ToString(Decimal value) => Convert.ToString(value);
 }
