+++ src/LuYao.Common/Valid/Valid.Int16.cs
@@ -5,32 +5,33 @@ namespace LuYao;
 partial class Valid
 {
     /// <inheritdoc/>
-    public static Int16 ToInt16(Boolean value) => value ? ((short)1) : ((short)0);
+    public static Boolean ToBoolean(Int16 value) => Convert.ToBoolean(value);
     /// <inheritdoc/>
-    public static Int16 ToInt16(Char value) => Convert.ToInt16(value);
+    public static Char ToChar(Int16 value) => Convert.ToChar(value);
     /// <inheritdoc/>
-    public static Int16 ToInt16(SByte value) => Convert.ToInt16(value);
+    public static SByte ToSByte(Int16 value) => Convert.ToSByte(value);
     /// <inheritdoc/>
-    public static Int16 ToInt16(Byte value) => Convert.ToInt16(value);
+    public static Byte ToByte(Int16 value) => Convert.ToByte(value);
     /// <inheritdoc/>
     public static Int16 ToInt16(Int16 value) => value;
     /// <inheritdoc/>
-    public static Int16 ToInt16(UInt16 value) => Convert.ToInt16(value);
+    public static UInt16 ToUInt16(Int16 value) => Convert.ToUInt16(value);
     /// <inheritdoc/>
-    public static Int16 ToInt16(Int32 value) => Convert.ToInt16(value);
+    public static Int32 ToInt32(Int16 value) => Convert.ToInt32(value);
     /// <inheritdoc/>
-    public static Int16 ToInt16(UInt32 value) => Convert.ToInt16(value);
+    public static UInt32 ToUInt32(Int16 value) => Convert.ToUInt32(value);
     /// <inheritdoc/>
-    public static Int16 ToInt16(Int64 value) => Convert.ToInt16(value);
+    public static Int64 ToInt64(Int16 value) => Convert.ToInt64(value);
     /// <inheritdoc/>
-    public static Int16 ToInt16(UInt64 value) => Convert.ToInt16(value);
+    public static UInt64 ToUInt64(Int16 value) => Convert.ToUInt64(value);
     /// <inheritdoc/>
-    public static Int16 ToInt16(Single value) => Convert.ToInt16(value);
+    public static Single ToSingle(Int16 value) => Convert.ToSingle(value);
     /// <inheritdoc/>
-    public static Int16 ToInt16(Double value) => Convert.ToInt16(value);
+    public static Double ToDouble(Int16 value) => Convert.ToDouble(value);
     /// <inheritdoc/>
-    public static Int16 ToInt16(Decimal value) => Convert.ToInt16(value);
+    public static Decimal ToDecimal(Int16 value) => Convert.ToDecimal(value);
     /// <inheritdoc/>
-    public static Int16 ToInt16(DateTime value) => Convert.ToInt16(value);
-
+    public static DateTime ToDateTime(Int16 value) => Convert.ToDateTime(value);
+    /// <inheritdoc/>
+    public static String ToString(Int16 value) => Convert.ToString(value);
 }
