+++ src/LuYao.Common/Valid/Valid.UInt16.cs
@@ -5,31 +5,33 @@ namespace LuYao;
 partial class Valid
 {
     /// <inheritdoc/>
-    public static UInt16 ToUInt16(Boolean value) => value ? ((ushort)1) : ((ushort)0);
+    public static Boolean ToBoolean(UInt16 value) => Convert.ToBoolean(value);
     /// <inheritdoc/>
-    public static UInt16 ToUInt16(Char value) => Convert.ToUInt16(value);
+    public static Char ToChar(UInt16 value) => Convert.ToChar(value);
     /// <inheritdoc/>
-    public static UInt16 ToUInt16(SByte value) => Convert.ToUInt16(value);
+    public static SByte ToSByte(UInt16 value) => Convert.ToSByte(value);
     /// <inheritdoc/>
-    public static UInt16 ToUInt16(Byte value) => Convert.ToUInt16(value);
+    public static Byte ToByte(UInt16 value) => Convert.ToByte(value);
     /// <inheritdoc/>
-    public static UInt16 ToUInt16(Int16 value) => Convert.ToUInt16(value);
+    public static Int16 ToInt16(UInt16 value) => Convert.ToInt16(value);
     /// <inheritdoc/>
     public static UInt16 ToUInt16(UInt16 value) => value;
     /// <inheritdoc/>
-    public static UInt16 ToUInt16(Int32 value) => Convert.ToUInt16(value);
+    public static Int32 ToInt32(UInt16 value) => Convert.ToInt32(value);
     /// <inheritdoc/>
-    public static UInt16 ToUInt16(UInt32 value) => Convert.ToUInt16(value);
+    public static UInt32 ToUInt32(UInt16 value) => Convert.ToUInt32(value);
     /// <inheritdoc/>
-    public static UInt16 ToUInt16(Int64 value) => Convert.ToUInt16(value);
+    public static Int64 ToInt64(UInt16 value) => Convert.ToInt64(value);
     /// <inheritdoc/>
-    public static UInt16 ToUInt16(UInt64 value) => Convert.ToUInt16(value);
+    public static UInt64 ToUInt64(UInt16 value) => Convert.ToUInt64(value);
     /// <inheritdoc/>
-    public static UInt16 ToUInt16(Single value) => Convert.ToUInt16(value);
+    public static Single ToSingle(UInt16 value) => Convert.ToSingle(value);
     /// <inheritdoc/>
-    public static UInt16 ToUInt16(Double value) => Convert.ToUInt16(value);
+    public static Double ToDouble(UInt16 value) => Convert.ToDouble(value);
     /// <inheritdoc/>
-    public static UInt16 ToUInt16(Decimal value) => Convert.ToUInt16(value);
+    public static Decimal ToDecimal(UInt16 value) => Convert.ToDecimal(value);
     /// <inheritdoc/>
-    public static UInt16 ToUInt16(DateTime value) => Convert.ToUInt16(value);
+    public static DateTime ToDateTime(UInt16 value) => Convert.ToDateTime(value);
+    /// <inheritdoc/>
+    public static String ToString(UInt16 value) => Convert.ToString(value);
 }
