+++ src/LuYao.Common/Valid/Valid.UInt32.cs
@@ -5,31 +5,33 @@ namespace LuYao;
 partial class Valid
 {
     /// <inheritdoc/>
-    public static UInt32 ToUInt32(Boolean value) => value ? 1u : 0u;
+    public static Boolean ToBoolean(UInt32 value) => Convert.ToBoolean(value);
     /// <inheritdoc/>
-    public static UInt32 ToUInt32(Char value) => Convert.ToUInt32(value);
+    public static Char ToChar(UInt32 value) => Convert.ToChar(value);
     /// <inheritdoc/>
-    public static UInt32 ToUInt32(SByte value) => Convert.ToUInt32(value);
+    public static SByte ToSByte(UInt32 value) => Convert.ToSByte(value);
     /// <inheritdoc/>
-    public static UInt32 ToUInt32(Byte value) => Convert.ToUInt32(value);
+    public static Byte ToByte(UInt32 value) => Convert.ToByte(value);
     /// <inheritdoc/>
-    public static UInt32 ToUInt32(Int16 value) => Convert.ToUInt32(value);
+    public static Int16 ToInt16(UInt32 value) => Convert.ToInt16(value);
     /// <inheritdoc/>
-    public static UInt32 ToUInt32(UInt16 value) => Convert.ToUInt32(value);
+    public static UInt16 ToUInt16(UInt32 value) => Convert.ToUInt16(value);
     /// <inheritdoc/>
-    public static UInt32 ToUInt32(Int32 value) => Convert.ToUInt32(value);
+    public static Int32 ToInt32(UInt32 value) => Convert.ToInt32(value);
     /// <inheritdoc/>
     public static UInt32 ToUInt32(UInt32 value) => value;
     /// <inheritdoc/>
-    public static UInt32 ToUInt32(Int64 value) => Convert.ToUInt32(value);
+    public static Int64 ToInt64(UInt32 value) => Convert.ToInt64(value);
     /// <inheritdoc/>
-    public static UInt32 ToUInt32(UInt64 value) => Convert.ToUInt32(value);
+    public static UInt64 ToUInt64(UInt32 value) => Convert.ToUInt64(value);
     /// <inheritdoc/>
-    public static UInt32 ToUInt32(Single value) => Convert.ToUInt32(value);
+    public static Single ToSingle(UInt32 value) => Convert.ToSingle(value);
     /// <inheritdoc/>
-    public static UInt32 ToUInt32(Double value) => Convert.ToUInt32(value);
+    public static Double ToDouble(UInt32 value) => Convert.ToDouble(value);
     /// <inheritdoc/>
-    public static UInt32 ToUInt32(Decimal value) => Convert.ToUInt32(value);
+    public static Decimal ToDecimal(UInt32 value) => Convert.ToDecimal(value);
     /// <inheritdoc/>
-    public static UInt32 ToUInt32(DateTime value) => Convert.ToUInt32(value);
+    public static DateTime ToDateTime(UInt32 value) => Convert.ToDateTime(value);
+    /// <inheritdoc/>
+    public static String ToString(UInt32 value) => Convert.ToString(value);
 }
